WIP experimenting with static constructor return types
