+++ crates/ink/src/codegen/dispatch/execution.rs
@@ -44,104 +44,39 @@ where
     Ok(())
 }
 
-/// Executes the given ink! constructor.
-///
-/// # Note
-///
-/// The closure is supposed to already contain all the arguments that the real
-/// constructor message requires and forwards them.
-#[inline]
-pub fn execute_constructor<Contract, F, R>(f: F) -> Result<(), DispatchError>
-where
-    Contract: Storable + StorageKey + ContractEnv,
-    F: FnOnce() -> R,
-    <private::Seal<R> as ConstructorReturnType<Contract>>::Error: Encode,
-    private::Seal<R>: ConstructorReturnType<Contract>,
-{
-    let result = ManuallyDrop::new(private::Seal(f()));
-    match result.as_result() {
-        Ok(contract) => {
-            // Constructor is infallible or is fallible but succeeded.
-            //
-            // This requires us to sync back the changes of the contract storage.
-            ink_env::set_contract_storage::<ink_primitives::Key, Contract>(
-                &Contract::KEY,
-                contract,
-            );
-            ink_env::return_value(ReturnFlags::default().set_reverted(false), &());
-        }
-        Err(error) => {
-            // Constructor is fallible and failed.
-            //
-            // We need to revert the state of the transaction.
-            ink_env::return_value::<
-                <private::Seal<R> as ConstructorReturnType<Contract>>::Error,
-            >(ReturnFlags::default().set_reverted(true), error)
-        }
-    }
-}
-
-mod private {
-    /// Seals the implementation of `ConstructorReturnType`.
-    pub trait Sealed {}
-    impl Sealed for () {}
-    impl<T, E> Sealed for Result<T, E> {}
-    /// A thin-wrapper type that automatically seals its inner type.
-    ///
-    /// Since it is private it can only be used from within this crate.
-    /// We need this type in order to properly seal the `ConstructorReturnType`
-    /// trait from unwanted external trait implementations.
-    #[repr(transparent)]
-    pub struct Seal<T>(pub T);
-    impl<T> Sealed for Seal<T> {}
-}
-
-/// Guards against using invalid contract initializer types.
-///
-/// # Note
-///
-/// Currently the only allowed types are `()` and `Result<(), E>`
-/// where `E` is some unspecified error type.
-/// If the contract initializer returns `Result::Err` the utility
-/// method that is used to initialize an ink! smart contract will
-/// revert the state of the contract instantiation.
-pub trait ConstructorReturnType<C>: private::Sealed {
-    /// Is `true` if `Self` is `Result<C, E>`.
-    const IS_RESULT: bool = false;
-
-    /// The error type of the constructor return type.
-    ///
-    /// # Note
-    ///
-    /// For infallible constructors this is `()` whereas for fallible
-    /// constructors this is the actual return error type. Since we only ever
-    /// return a value in case of `Result::Err` the `Result::Ok` value type
-    /// does not matter.
-    type Error;
-
-    /// Converts the return value into a `Result` instance.
-    ///
-    /// # Note
-    ///
-    /// For infallible constructor returns this always yields `Ok`.
-    fn as_result(&self) -> Result<&C, &Self::Error>;
-}
-
-impl<C> ConstructorReturnType<C> for private::Seal<C> {
-    type Error = ();
-
-    #[inline(always)]
-    fn as_result(&self) -> Result<&C, &Self::Error> {
-        Ok(&self.0)
-    }
-}
-
-impl<C, E> ConstructorReturnType<C> for private::Seal<Result<C, E>> {
-    const IS_RESULT: bool = true;
-    type Error = E;
-
-    #[inline(always)]
-    fn as_result(&self) -> Result<&C, &Self::Error> {
-        self.0.as_ref()
-    }
-}
+// /// Executes the given ink! constructor.
+// ///
+// /// # Note
+// ///
+// /// The closure is supposed to already contain all the arguments that the real
+// /// constructor message requires and forwards them.
+// #[inline]
+// pub fn execute_constructor<Contract, F, R>(f: F) -> Result<(), DispatchError>
+// where
+//     Contract: Storable + StorageKey + ContractEnv,
+//     F: FnOnce() -> R,
+//     <private::Seal<R> as ConstructorReturnType<Contract>>::Error: Encode,
+//     private::Seal<R>: ConstructorReturnType<Contract>,
+// {
+//     let result = ManuallyDrop::new(private::Seal(f()));
+//     match result.as_result() {
+//         Ok(contract) => {
+//             // Constructor is infallible or is fallible but succeeded.
+//             //
+//             // This requires us to sync back the changes of the contract storage.
+//             ink_env::set_contract_storage::<ink_primitives::Key, Contract>(
+//                 &Contract::KEY,
+//                 contract,
+//             );
+//             ink_env::return_value(ReturnFlags::default().set_reverted(false), &());
+//         }
+//         Err(error) => {
+//             // Constructor is fallible and failed.
+//             //
+//             // We need to revert the state of the transaction.
+//             ink_env::return_value::<
+//                 <private::Seal<R> as ConstructorReturnType<Contract>>::Error,
+//             >(ReturnFlags::default().set_reverted(true), error)
+//         }
+//     }
+// }
