+++ crates/ink/src/codegen/dispatch/mod.rs
@@ -19,7 +19,7 @@ mod type_check;
 pub use self::{
     execution::{
         deny_payment,
-        execute_constructor,
+        // execute_constructor,
     },
     info::ContractCallBuilder,
     type_check::{
